+++ tests/test_objects.py
@@ -4,7 +4,7 @@ import decimal
 import re
 
 import pydash as pyd
-from pydash.api.utilities import _iterate
+from pydash.api.utilities import iterator
 
 from . import fixtures
 from .fixtures import parametrize
@@ -39,7 +39,7 @@ def test_clone(case, args):
 
     assert actual is not case
 
-    for key, value in _iterate(actual):
+    for key, value in iterator(actual):
         assert value is case[key]
 
 
@@ -56,7 +56,7 @@ def test_clone_deep(case, kargs):
     for actual in actuals:
         assert actual is not case
 
-        for key, value in _iterate(actual):
+        for key, value in iterator(actual):
             assert value is not case[key]